CHAT!!!!!!

+++ src/main/java/com/miriamsonaglia/mediasharehouse/service/HouseManager.java
@@ -117,45 +117,53 @@ public class HouseManager {
     public void deleteHouse(int idCasa) {
         try (Connection connection = DatabaseConnection.getConnection()) {
             connection.setAutoCommit(false);
-
+    
             CasaDao casaDao = new CasaDao(connection);
             StanzaDao stanzaDao = new StanzaDao(connection);
             ContenutoDao contenutoDao = new ContenutoDao(connection);
-
-            // Ottieni tutte le stanze della casa
-            List<Stanza> stanze = stanzaDao.getStanzeByCasa(idCasa);
-
-            // Elimina i contenuti di ogni stanza
-            for (Stanza stanza : stanze) {
-                List<Contenuto> contenuti = contenutoDao.getContenutiByStanza(stanza.getIdStanza());
-                for (Contenuto contenuto : contenuti) {
-                    contenutoDao.deleteContenuto(contenuto.getIdContenuto());
+    
+            // Conferma l'eliminazione
+            int confirm = JOptionPane.showConfirmDialog(frame, "Sei sicuro di voler eliminare questa casa e tutte le sue stanze e contenuti?", "Conferma eliminazione", JOptionPane.YES_NO_OPTION);
+            if (confirm == JOptionPane.YES_OPTION) {
+                // Ottieni tutte le stanze della casa
+                List<Stanza> stanze = stanzaDao.getStanzeByCasa(idCasa);
+    
+                // Elimina i contenuti di ogni stanza
+                for (Stanza stanza : stanze) {
+                    List<Contenuto> contenuti = contenutoDao.getContenutiByStanza(stanza.getIdStanza());
+                    for (Contenuto contenuto : contenuti) {
+                        contenutoDao.deleteContenuto(contenuto.getIdContenuto());
+                    }
+                    // Elimina la stanza
+                    stanzaDao.deleteStanza(stanza.getIdStanza());
                 }
-                // Elimina la stanza
-                stanzaDao.deleteStanza(stanza.getIdStanza());
-            }
-
-            // Elimina la casa
-            boolean isDeleted = casaDao.deleteCasa(idCasa);
-
-            if (isDeleted) {
-                connection.commit();
-                JOptionPane.showMessageDialog(frame, "Casa eliminata con successo.");
-                MSHHome.removeHouseButtonFromPanel(idCasa);
+    
+                // Elimina la casa
+                boolean isDeleted = casaDao.deleteCasa(idCasa);
+    
+                if (isDeleted) {
+                    connection.commit();
+                    JOptionPane.showMessageDialog(frame, "Casa eliminata con successo.");
+                    MSHHome.removeHouseButtonFromPanel(idCasa);
+                } else {
+                    connection.rollback();
+                    JOptionPane.showMessageDialog(frame, "Errore durante l'eliminazione della casa.");
+                }
             } else {
                 connection.rollback();
-                JOptionPane.showMessageDialog(frame, "Errore durante l'eliminazione della casa.");
             }
         } catch (SQLException ex) {
             ex.printStackTrace();
-            /*try {
+            try {
                 // Rollback in caso di errore
-                connection.rollback();
+                DatabaseConnection.getConnection().rollback();
             } catch (SQLException rollbackEx) {
                 rollbackEx.printStackTrace();
-            }*/
+            }
+            JOptionPane.showMessageDialog(frame, "Errore nella connessione al database.");
         }
     }
+    
 
     public void removeAccessToHouse(int idCasa) {
         try (Connection connection = DatabaseConnection.getConnection()) {
